Modify `Publisher.first()` to not emit error (#437)
__Motivation__

`Publisher.first()` today either emits the first item or `NoSuchElementException` if there were no items emitted.
So technically, it is `firstOrError`, but we already have `firstOrError` that emits error for more scenarios.
We can modify `first()` to not emit an error when the `Publisher` is empty, instead use a default value. This will make the names clearer.

__Modification__

Take a `Supplier<T>` for `first()` and use it to get a value when there is none emitted from the source `Publisher`.

__Result__

`first()` now does not emit a error and hence is true to its name.