Qualify `flatMap` with "merge" where applicable
__Motivation__

Flattening a result can be done in order or in parallel. For `Publisher#flatMapSingle` variants we process each new `Single` in parallel so there is no order guarantee between the `Single` created for an emitted item and the result of that single as emitted on the final stream. eg:

`from(1,2,3).flatMapSingle(t -> aSingleForT)`

Assuming each generated `aSingleForT` generates items `T1`, `T2`, `T3` for the original items `1, 2, 3` then the values `T1`, `T2`, `T3` can be emitted in any order from the resulting `Publisher`. This behavior is subtle and is currently misrepresented in the name which could lead to confusion.

__Modification__

- Rename `flatMapSingle*` to `flatMapMergeSingle*`
- Rename `concatMapIterable` to `flatMapConcatIterable`
- Removed `flatMapIterable` as it just uses `concatMapIterable` and there is just one usage of it. We can do without this operator in the public API.

__Result__

More intuitive names.