Provide a way for server builder to specify an execution strategy
__Motivation__

We thought `*HttpService` providing a method to provide it’s strategy is more useful as the service knows the offloading requirement, however, this means that we need to update the strategy provided by the service if we find a conflicting filter in the filter chain. This requirement makes it such that the strategy handling is different for client and server; as at the client we never merge strategies for provided at the builder. Understanding execution strategy behavior is complex in itself and having different behaviors for client and server makes it harder to understand.

We should make the behavior/expectations from user consistent and hence we will have the user specify strategy at the server builder understanding the filter chain and if a custom strategy is provided, ST will never override it.

__Modification__

- Added `executionStrategy()` method for the server builder.
- Removed `executionStrategy()` method for the service filters.
- Fixed wrongly specified default strategies for aggregated programming models.

__Result__

Consistent behavior across client and server with user specified custom strategy.