HttpConnection limiters should use onClosing instead of onClosed
Motivation:
The ReservableRequestConcurrencyController serves to limit connections when the
underlying transport cannot handle requests. Currently
AbstractReservableRequestConcurrencyController triggers of of the settings
stream for max concurrent requests and the connection onClose method. The
onClose will only trigger when the underlying transport is fully closed. This
may mean that requests are allowed on the connection in a partially closed state
of which requests are unlikely to succeed.

Modifications:
- AbstractReservableRequestConcurrencyController should override and avoid
  offloading because it subscribes internally and doesn't block
- Client and Connection usages of AbstractReservableRequestConcurrencyController
should pass onClosing() instead of onClosed() to stop requests on the leading
edge of shutdown and half closure

Result:
Earlier detection of unusable connection and less failed requests from the
client's perspective.