Remove do* prefix on operators (#473)
Motivation:
The do* is sometimes used inconsistently relative to the do<method name>. For
example there are operators such as Publisher#doOnCancel() and
Publisher#doOnNext. In the first case the "doOn" is meant to highlight 'the
callback will be invoked on the cancel method call' and in the second case it
means 'the callback is invoked when the onNext method is call'. This leads to
issues when considering operators such as "doOnSubscribe" and "doSubscribe". The
second method name implies some subscribe action will be performed, but this is
not the case.

Modifications:
- [doOn | do]<method name> -> when<method name>
- doBefore<method name> -> before<method name>
- doAfter<method name> -> after<method name>

Result:
Less conflicts in the naming approach for methods that don't modify the
underlying stream and may have side effects.