Setting `content-length` for aggregated requests/responses.
Motivation:

For requests/responses where we know the payload is all available in
memory, sending a `content-length` header offers better performance in a
number of cases.

Modifications:

- Add optional `EffectiveApiType` to request/response classes.
  - Carried through most conversions.
  - Adding/transforming trailers adjusts flag appropriately.
- Add either `content-length` or `transfer-encoding: chunked` prior to
  sending request/response.
  - Omit `content-length` in cases where the RFC indicates this is allowed.

Results:

Requests/responses created via the aggregated API will generally be sent
with a `content-length` header, while all others will be sent with
`transfer-encoding: chunked` as was done previously.