Optimize initial line parsing in HttpResponseDecoder
Motivation:

`HttpResponseDecoder` creates 2 slices of the Netty's `ByteBuf` when it
parses the initial line. HTTP status code could be parsed without
slicing at all. The status reason phrase is relatively small and it is
cheaper to convert it to `String` to avoid slicing and utilize JDK
optimizations for `String`.

Modifications:

- Implement optimized version of decoding HTTP status code, which
does not require memory allocation;
- Throw in exception if the status code is not a 3-digit integer;
- Use `ByteBuf.toString` instead of `ByteBuf.slice` in
`HttpResponseDecoder` for HTTP status reason phrase;
- Remove unused `HttpResponseStatus.of(int, Buffer)` factory method;
- Use `PREFER_HEAP_RO_ALLOCATOR` instad of `PREFER_DIRECT_RO_ALLOCATOR`
in `HttpResponseStatus` to allocate intenal `Buffer`s;
- Add `HttpResponseDecoderBenchmark` and `HttpResponseEncoderBenchmark`;
- Update tests;

Result:

Improved performance of `HttpResponseDecoder` and `HttpResponseEncoder`.