Add a BasicAuth JAX-RS security filter that works with BasicAuthHttpServiceBuilder
__Motivation__

`BasicAuthHttpServiceBuilder` can be used to create filters that secure ServiceTalk requests with Basic authentication.

The security of JAX-RS applications is based on the association of `SecurityContext` object with the in-flight requests.

Thus we need to offer support for setting `SecurityContext` for requests that have been Basic-authenticated at ServiceTalk level.

__Modifications__

Add JAX-RS security filters that work with the global and name binding strategies.

__Results__

JAX-RS applications can benefit from the security provided by `BasicAuthHttpServiceBuilder`.