RedisRequesterUtils may allocate more memory than necessary
Motivation:

When we aggregate `BulkStringChunk`s in `RedisRequesterUtils` we are
making sure that `aggregator` will be able to aggregate all chunks.
However, when we calculate expected bytes for the next chunks we do not
subtract the number of bytes of the `FirstBulkStringChunk` we already
have written to this `aggregator`.

Modifications:

- Subtract the size of existing `Buffer` when calculating expected bytes
in `RedisRequesterUtils`;
- Move calculation logic to a separate method;

Result:

`RedisRequesterUtils` allocates only required number of bytes for the
aggregator.