Reduce number of wrapping layers for `ReadOnlyBuffer`
Motivation:

1. `ReadOnlyBuffer` controls read-only access and there is no need to
convert internal `Buffer` to read-only when we create a new instance.
2. It's not necessary to wrap the internal `Buffer` with additional
`ReadOnlyBuffer` layer when we slice or duplicate if the internal
`Buffer` is already read-only.
3. It's not necessary to wrap the `ByteBuffer` with additional
read-only layer when we convert `toNioBuffer` if the internal
one is already read-only.

Modifications:

- Remove `preserveNettyInvariants`;
- Use `asReadOnly()` instead of `new ReadOnlyBuffer` which will not wrap
if not necessary;
- Check that `ByteBuffer` is not already a read-only before invoking
`ByteBuffer.asReadOnlyBuffer()` which will save `duplicate()` call;

Result:

Less wrapping layers and allocations for `ReadOnlyBuffer`.