HttpCookiePair and HttpSetCookie parsing premature name matching
Motivation:
The parsing logic for HttpCookiePair and HttpSetCookie would match if the
nprovided name shared a common prefix. This will lead to matching of incorrect
matching and removal of cookies whose names don't actually match.

Modifications:
- When matching the name for HttpCookiePair and HttpSetCookie, we must parse
until the '=' token to be sure we have found the end of the name.
- Share HttpCookiePair parsing code which must parse a cookie-string that may
contain multiple cookie-pairs
https://tools.ietf.org/html/rfc6265#section-4.2.1.

Result:
HttpCookiePair and HttpSetCookie parsing no longer matches the search name as a
subsequence.