Reduce contention when Round Robin (#535)
* Reduce contention when Round Robin

__Motivation__

When the number of addresses is small and connections per address is large, the current RRLB implementation keeps attempting to select from the front of the queue. This results in many needlessly failing CAS operations before finding the available connection. This problem is amplified as the number of connections increases and starts to consume considerable CPU time above 100 connections.

__Modifications__

Change RRLB to leverage a Copy On Write, datastructure to allow for a stable randomly accessible set when selecting connections. The algorithm uses random selection in a snapshot of the set to find an available connection up to number of attempts, defined by the search factor.

__Result__

Lower CPU overhead while selection connections, much higher throughput overall at similar tail latencies.