Improve payload draining performance
__Motivation__

ST automatically drains the payload when a user forgets to subscribe but this result in duplicate subscriptions to the payload on every request when the user does do the right thing or uses an API which does the same (Aggregated Streaming and Blocking). ST should detect a subscription took place and only drain when that's not the case.

__Modifications__

- Set a flag when the payload is subscribed to
- Upon response completion check flag and only condionally drain the payload

__Result__

No more DuplicateSubscriotionExceptions thrown on the hot and happy path