Reset thread interrupt status when catching `InterruptedException`
__Motivation__

When an `InterruptedException` is thrown, the thread's interrupted status is reset.
If we are re-throwing and not handling interrupts, then we should set the status so that any subsequent blocking code can see the interrupt status.

__Modification__

Reset the interrupt status by calling `Thread.currentThread().interrupt();` where we are catching `InterruptedException`

__Result__

Better handling of interrupts.