Validate user supplied values in request/response
__Motivation__

Invalid values like `null` passed as headers, cookies, query parameters must be validated. Failure to do so leads to hard to debug issues, eg: today if header value is `null` the client appears to hang as we throw an error from the HTTP encoder which gets unnoticed and nothing gets written to the channel (if there was no payload, trailers). This leads to hard to debug issues specifically if such invalid values are passed conditionally.

__Modification__

- Make sure we are validating such user input for metadata.
- Made the validations consistent (always throw `IllegalArgumentException`

__Result__

Better validations and early failures in presence of invalid values.