Use a NOOP resolver on Netty's Bootstrap to avoid caching EventLoopGroups
Motivation:
Netty will attempt to resolve addresses when connecting via the Bootstrap. In
this process Netty will attempt to get an AddressResolver and
AddressResolverGroup will cache the result in an IdentityMap keyed by
EventExecutor. The entry will be removed from the map when the
terminationFuture() of the EventExecutor completes, but it isn't necessary to
insert in the map in the first place because the ServiceDiscoverer is
responsible for resolving addresses.

Modifications:
- Create NoopNettyAddressResolverGroup which is an AddressResolverGroup that
doesn't cache anything, and always returns success for all resolution attempts.

Result:
Less duplication of logic in Netty and unnecessary retention of EventExecutor
objects.