Retrying HTTP requests when first write fails. (#582)
Motivation:

When a connection is closed unexpectedly after a response (eg. in HTTP/1.1
without a `connection: close` header) the load balancer may not detect this
until another read or write operation is performed on the request. This
means that subsequent requests may occasionally fail due to the connection
having been previously closed.

Modifications:

Surface a retryable exception when the first write operation in a stream
fails due to the connection being closed.
Move `RetryableException` to the transport-api module.

Results:

The default retry filter retries when a previously closed connection is
re-used for a new request.