Remove `HttpConnectionBuilder`
__Motivation__

`HttpConnectionBuilder` is not an easy to use abstraction as it does not provide any connection pooling capabilities out of the box and reusing connections along with considering connection closures is difficult to do by the user.
In addition maintaining this as a public API has maintenance and API compatibility overhead.

__Modification__

- Added `HttpClients.forResolvedAddress()` variants to make it easy for users to replace usage of connection builder.
- Replace test usages of `HttpConnectionBuilder` with `reserveConnection()` as suggested for users.
- Removed builder related methods from `DefaultHttpConnectionBuilder` and made it a static connection factory.

__Result__

Less public API, less maintenance overhead.