Use flushOnEnd for aggregated HTTP API.
Motivation:

When using the aggregated HTTP API, we can safely flush at the end of the
payload (or whenever Netty's buffers are full) rather than after the
metadata and each payload item. This reduces the number of system calls,
improving performance.

Modifications:

- Pass the `FlushStrategy` through a few layers of internal APIs.
- Detect when the aggregated API is being used, and set `flushOnEnd`.

Results:

The number of system calls for flushes are reduced.