Updating flush strategy should know if the strategy is already updated.
__Motivation__

`NettyConnectionContext#updateFlushStrategy()` provides the current `FlushStrategy` but there is no way of knowing if the current strategy is being already updated.
This makes it hard to implement cases where someone wants to apply a strategy only if there isn't a strategy already provided by the user. Such cases are typically useful for dynamic defaults, eg: provide a strategy based on programming model.

__Modification__

Modify the current `updateFlushStrategy()` method to now also provide a flag if the passed `current` `FlushStrategy` is the original strategy for the connection.

__Result__

Providing dynamic defaults is possible.