Do not create unresolved `InetSocketAddress` for resolved address clients.
__Motivation__

`HttpClients.forResolvedAddress()` methods create an `unresolvedAddress` if we are provided with an IP address.
With the recent changes of not doing any resolution while connection, makes it such that we can not use this method.

__Modification__

Use the regular `InetSocketAddress` constructor. If the address is actually resolved, it will not do the resolution, which is what we want.

__Result__

`HttpClients.forResolvedAddress()` is usable.