Fix flaky testConsumeRequestPayloadAfterTrailersSent
Motivation:

The `testConsumeRequestPayloadAfterTrailersSent` test from
`ConsumeRequestPayloadOnResponsePathTest` had concurrency issue. Because
the client receives the full response in this test `try-with-resources`
block will close the client which leads to closing the connection and
canceling the `Publisher` of a payload body. It triggers `afterFinally`
in a different thread, counts down the latch, and starts the test
assertions. We should not close the client if the server didn't complete
the request payload body processing.

Modifications:

- await on the server latch inside `try-with-resources` block;
- rename `CountDownLatch` variable: waitServer -> serverLatch;

Result:

Resolves #414.