Log plaintext of SSL connection on server side. (#619)
Motivation:

When enabling wire logging, users generally want to see protocol level
plaintext output, not the encrypted form. The client side already logs
this way, but the server side was logging the encrypted form.

Modification:

Add the wire logger at the end of the channel initialization, on the
server.

Results:

Wire logs show plaintext for both client and server.