Create an empty heap Buffer instead of direct one if payload body is null
Motivation:

If ST receives a request or response with an empty payload body,
aggregation logic allocates a new `Buffer` of size `0` (instead of using
an `EMPTY_BUFFER` constant) for users in case they need to expand it.
Allocation of direct memory is very expensive.

Modifications:

- Allocate an empty heap `Buffer` in
`HttpDataSourceTransformations.aggregatePayloadAndTrailers` instead of
direct one;

Result:

Improved throughput for aggregated API by 8-16%, depending on the use
case.