Use `EmptyHttpHeaders` if decoded request/response may not have trailers (#630)
Motivation:

`HttpHeaders` object is used as a marker of the end of request or
response object for `HttpObjectEncoder/HttpObjectDecoder`. If ST does
not expect trailers according to the HTTP protocol, it will ignore this object.
For these situations we don't need to allocate a new `DefaultHttpHeaders`
object.

Modifications:

- Use `EmptyHttpHeaders.INSTANCE` instead of
`headersFactory.newEmptyTrailers()` where possible;

Result:

Less allocation of unused objects, simplified API of internal methods.