FlushStrategyProvider#getNewStrategy(..) rename clarify intent
Motivation:
We generally don't use the "get" prefix on method names and it is somewhat
ambiguous on a factory/provider interface. The "new" terminology may also be
confusing if implementers assume they have to generate a "new" object to isolate
state.

Modifications:
- Rename FlushStrategyProvider#getNewStrategy(..) to
FlushStrategyProvider#computeFlushStrategy(..)

Result:
FlushStrategyProvider method name is clarified and more consistent.