SingleFlatMapPublisher to use CancellableThenSubscription intead of SequentialSubscription (#641)
Motivation:
SingleFlatMapPublisher currently uses SequentialSubscription to orchestrate the
Subscription hand off between the Single and Publisher. However this is overkill
as we don't need to manage two full Subscriptions but instead a Cancellable then
a Subscription.

Modifications:
- Use the recently introduced CancellableThenSubscription class in
SingleFlatMapPublisher instead of SequentialSubscription.

Result:
Lower overhead management of the Subscription in SingleFlatMapPublisher.