Make predicate router support smooth evolution to non-blocking (#648)
__Motivation__

When an HTTP server and router is built the offloading is computing in 2 places.

1. on the HttpServerBuilder
2. on each route

The current implementation makes it challenging to evolve an application from blocking to non-blocking. For example if someone wants a route that is non-blocking they have to change the offload strategy on (1) which impacts all routes and may impact code beyond what the user is responsible for. This high bar makes it unlikely users will be able to make effective use of the mechnims. We should instead have a mechansim where the user can evolve the strategy on individual routes (e.g. like on the client side) so users can evolve in a controlled manner.

__Modifications__

- Fix predicate router to support smooth evolution to non-blocking
- Improve Jersey router computation of effective offloading strategy
- Increase predicate router test coverage

__Results__

- Users of the predicate router can transition to non-blocking progressively, i.e. in a route by route manner.
- The Jersey router is ready for a more accurate computation of effective offloading strategy