Handling write failures that do not close the channel (#585)
__Motivation__

When writing items from a `Publisher` on a channel, we assume that any write failures will result in channel closure as write failures are not recoverable on an ordered protocol. This led to the optimization that we do not need to add listeners for each writes but we can do so for the last seen write and the result of the last write will be the result of all previous writes.
This approach works great and optimally when our basic assumption (write failures will close the channel) but in practice we have handlers in the pipeline that do not protect against unexpected failures (eg: throw `Exception` from `write()`). Netty only fails the write promise if `Handler#write()` fails and if there are no listeners to that promise, the failure is lost. This leads to hung connections and hard to debug situations.

__Modification__

Modified `WriteStreamSubscriber` to use a special `ChannelPromise` for all writes originated from the same `Publisher`.
This special promise does the following:

- Close channel on first write failure.
- Counts writes and completions, when all writes are done and original `Publisher` has terminated, terminate the associated `Completable`.
- Inform listeners (for all writes, if any) only when all writes and the original `Publisher` has terminated.

The trade-off here is that we do not use a new promise for each write which means any listeners added to any interim writes will not be informed till all writes are done. In practice we do not add listeners to intermediate writes anyways, so the trade-off makes sense.

__Result__

Better visibility of intermediate write errors in a stream.