Scripts for managing tags in antora-remotes.yml
Motivation:

In order to ensure that antora docs contain all release versions, we need
to add each the release tag to `site-remote.yml` when we release a major or
minor version.
In order to ensure each release builds docs from the latest patch version,
we need to update the release tags in `site-remote.yml` when we release
a patch version.

Modifications:

Add scripts for managing the tags.

Results:

We can easily manage the tags.