Global scheduler is unsafe to use (#654)
__Motivation__

If a user creates an `Executor` with a non-scheduled `ExecutorService`, we use a global scheduler to schedule timers.
All timer ticks then invoke user code in the global scheduler thread. If user code blocks, we will backup the global scheduler and delay further timers.

__Modification__

For cases when user does not provide a `ScheduledExecutorService`, we will use a single global scheduler and offload timer ticks to the user provided executor.
This will not add more threads in the system and still be safe in presence of blocking code.

__Result__

Global scheduler is safe to use in presence of blocking code.