Fix inferring Content-Length for HEAD responses
__Motivation__

HEAD requests should either have the content-length already set (= what GET will return) or
have the header omitted when unknown, but never have any payload anyway so we shouldn't try to infer it.
This lead to a bug when folks use the aggregated API the current code would infer to length 0 always,
which is incorrect when no payload was set.

__Modifications__

Ensure that we don't try to compute Content-Length for HEAD responses
Update tests to reflect the expected behavior.

__Result__

Aggregation now either retains the pre-existing content-length or omits computing the header.