Use pooled buffers for reading from the socket (#656)
Motivation:

Transport layer requires direct buffers for reading from the socket.
Frequent allocation of direct memory is very expensive in JVM when an
application hits the direct memory limit, because JVM will trigger full
GC before reserving a new direct memory chunk. Heap memory allocation is
cheaper. We may utilize pooled direct buffers when we read data from the
socket and copy them to the heap memory before handed over to the user
to make sure users not need to care about releasing buffers.

Modifications:

- Introduce `PooledRecvByteBufAllocatorInitializer` that sets
`ThreadLocalRecvByteBufAllocator` for the `Channel` and adds
`CopyByteBufHandler` as the first handler of the `Channel` pipeline;
- Utilize this `PooledRecvByteBufAllocatorInitializer` in
`TcpClientChannelInitializer` and `TcpServerChannelInitializer`;

Result:

Improved performance for the client and server:
  +23% RPS for reading requests without payload body;
  +38% RPS for reading requests with 256B payload body;
  +64% RPS for reading requests with 16K payload body;
  +GC overhead dropped from ~13% of CPU time to <1%.