Enable TCP fallback and notify once preferred records are received in service discovery
Motivation:

Since netty 4.1.37.Final we support TCP fallback if a truncated dns response is received (due datagram limits). We should use this.
Beside this we also now support to notify as soon as all preferred records are received which is good enough for the service discovery. Let's enable it.

Modifications:

- Specify SocketChannel type for TCP fallback
- enable fast notify

Result:

More robust and fast service discovery