Add `SslProvidersTest` for larger payload body
Motivation:

It was discovered that ordering of channel handlers
(`CopyByteBufHandler` -> `SslHandler`) matters for the larger payload
size, 16KB+. With the opposite ordering, request fails with
`SSLException` when using `SslProvider.JDK`.

Modifications:

- Add tests to verify that `SslHandler` works correctly with larger
payload body;
- Move `tcnativeVersion` to the `servicetalk-bom-internal` to keep the
version number close to the netty's version;

Result:

Tests to verify that SSL works for larger payload body with all the
handlers we set on the channel pipeline.