Linking to correct source code version from the docs.
Motivation:

Previously, release docs attempted to link to a git tag that was missing the
patch version, which would have let to broken links. We need to link to the
correct git tag, which includes the patch vesion.

Modifications:

Add `git-tag` attribute to antora modules, and use it for linking to source
for releases.
Add a script for the release process that updates this attribute.

Results:

Generated release docs link to correct git tag.