Fixing doc generation by downloading and committing ui-bundle.zip
Motivation:

Doc generation was broken because the `ui-default` underwent some structural
changes that are incompatible with our `supplemental-ui` files.
As a quick-fix, I've commited the `ui-bundle.zip` so we're not dependent
on the `bundle-stable`  build.

Modifications:

Commit `ui-bundle.zip` and reference it locally.

Results:

Our doc build run and generates docs repeatably.