Publisher#concat(Single) to propagate an error if onNext throws (#673)
Motivation:
PublisherConcatWithSingle delivers signals to the downstream Subscriber from a
terminal method in SingleSource.Subscriber#onSuccess. If the downstream
Subscriber throws from onNext then the error may never be propagated.

Modifications:
- PublisherConcatWithSingle should catch any exceptions from onNext when
attempting to terminate the downstream Subscriber when the Single completes.
This can also be done from the Subscription thread out of convenience because
the concurrency protection already exists and it provides better exception
propagation (as opposed to throwing in the calling thread).

Result:
More robust error handling in Publisher#concat(Single).