`WriteStreamSubscriber` may use a completed promise for write
__Motivation__

`WriteStreamSubscriber` reuses the same promise instance for all writes through a `Subscriber` instance.
It may happen that we get more `onNext`s after the channel is already closed. This leads to sending a completed promise for a write.
Netty will try to fail the promise in such case (as the channel is closed) and will log that exception. Although the write has to fail anyways and we have already terminated our control flow, the exception is benign but it can be noisy.

__Modification__

After we have terminated our control flow or if the channel is closed, ignore any subsequent `onNext`s.

__Result__

No noise for write failures after channel is closed.