Some improvements to the performance documentation (#693)
__Motivation__

Our intention for writing the performance documentation is to share various trade offs that ServiceTalk has made which are interesting from a performance point of view. Sharing the internal performance evaluation tooling details do not add value to this message and has the potential to distract from the message.
Till we are confident about sharing the internal tooling publicly, it is better to focus on this message and keep the tooling details for a separate document.

__Modification__

As discussed here: https://github.com/apple/servicetalk/pull/684#discussion_r303177955 remove the sections which are not yet required to be shared.

__Result__

More focused message in our documents.