Publisher#flatMapConcatIterable not catching exceptions when delivering data from Subscription thread
Motivation:
Publisher#flatMapConcatIterable may deliver data from the Subscription thread.
In this case it is the source of data and should catch exceptions and propagate
them to the Subscriber. Otherwise it is possible exceptions will not propagate
as expected.

Modifications:
- PublisherConcatMapIterable should catch and propagate exceptions from the
Iterator/Subscriber when delivering data from the Subscription thread.

Result:
More correct error handling in Publisher#flatMapConcatIterable