Optimistically propagate exceptions instead of log/swallow (#690)
Motivation:
The ReactiveStreams specification states that methods should generally not throw
exceptions, and instead propagate errors via Subscriber#onError. Since we
implement operators which may invoke user code we can't be sure if users will
follow these rules, or even be aware of them, so we defensively catch
exceptions. It is not always clear how to react when onSubscribe method throws
as the downstream Subscriber chain may (or may not) have been properly
initialized. In some instances we were logging the exception and not propagating
the control flow to avoid violating other clauses of the ReactiveStreams spec
[2][3]. Not propagating the exception may result in the asynchronous control
flow not completing which causes a timeout/hang.

[1] https://github.com/reactive-streams/reactive-streams-jvm#1.9
[2] https://github.com/reactive-streams/reactive-streams-jvm#1.7
[3] https://github.com/reactive-streams/reactive-streams-jvm#1.3

Modifications:
- Instead of logging exceptions from onSubscribe, we should propagate them to
onError.
- Change log level from "debug" to "info" when unexpected exceptions are caught
and logged on terminal Subscriber methods.

Result:
Less likely to swallow exceptions and instead make a best effort to complete
asynchronous control flow by propagating errors.