Remove APIs for unsupported features.
__Motivation__

Compression and deadline are not yet supported but we have them in the API.
We do not intend to support these features in the near term so we should not allow users to use APIs that are not supported.

__Modification__

Remove these methods from the APIs.

__Result__

APIs that are not going to be implemented in the near future are not exposed to users.