Implement Jersey Router in all programming models (#706)
__Motivation__

Currently the programming model defines which optimizations are used wrt offloading and flushing.
Before this change there was only a Jersey router implemented on the `StreamingHttpService` API
which results in offloading all signals and flushing to the transport for every element of data.
Jersey applications are typically BlockingAggregated or BlockingStreaming in nature, which means
that we can reduce to a single offload of the user code and in case of aggregated a single flush.
These optimizations result in a significant boost in performance for typical Jersey applications,
without sacrificing the ability for users to leverage Async sources such as Single, Publisher,
Completable and the user had the ability to opt for the model that most closely reflects their
use-case to get the best performance. This comes with the drawback that users expecting streaming when opting for an aggregated router, that aggregation will happen behind the scenes such that data will not be streaming in nature. For now we accept this drawback until we have better APIs in place to automagically infer the best offloading and flushing strategy based on the route.

__Modifications__

- Implement API conversion from the streaming Jersey to AsyncAggregated, BlockingAggregated and
BlockingStreaming
- Update Jersey tests to run under all programming models to ensure backwards compatibility

__Result__

Typical Jersey applications may benefit up to 55% (BA) and 22% (BS) performance boost over the current Streaming router when offloading is enabled. For BA offloading will need to be enhanced in the future to reach that level of performance, it's currently slower than Streaming because of unnecessary offloading.