Allow trailer updates for streaming request/response when payload emits error
__Motivation__

We have `transform()` methods on streaming request/response that allows for trailers to be updated when the payload `Publisher` terminates successfully.
There are scenarios (gRPC) where we even have to update the trailers for errors.

__Modification__

Introduce a `TrailerTransformer` interface to expose the different facets required for `transform` methods.
`TrailerTransformer` also has a callback to update trailers on error.

__Result__

Trailers can be updated even when payload emits an error.