Restructure SSL/TLS configuration (#735)
__Motivation__

Current way of configuring security is not suitable for layered protocol builders.
gRPC uses HTTP builders and would be ideal if we can reuse abstractions. SSL config builders classes being `final` does not help.
We expose `SSLConfig` as well as the builder to build `SSLConfig`. Since, `SSLConfig` isn't expected to be used by users, we can internalize the config and only expose the builders to the users. This will reduce API surface area and simplify the cognitive overhead for users to understand the various layers.

__Modification__

- Introduced `*SecurityConfigurator` as the user facing API.
- Introduced internal `SecurityConfig` to be used by the protocols, similar to protocol builder configurations we use elsewhere.
- Removed `SslConfig` and the related builders.

__Result__

Easier and composable security configuration.