MultiAddressHttpClient should accept only absolute-form request-target (#720)
Motivation:

Like in many other HTTP clients, it's more easier and convenient for
users to use absolute-form request-target (i.e. full URL) when they are
using multi-address client. It also helps to reduce configuration for
the client, eliminating `effectiveScheme` function from the builder.

Modifications:

- Remove unnecessary `MultiAddressHttpClientBuilder.effectiveScheme`;
- `MultiAddressHttpClient` throws if the request-target is not in
absolute-form;
- `MultiAddressHttpClient` transforms request-target from absolute-form
to relative-form;
- `RedirectSingle` should support new behavior of
`MultiAddressHttpClient` to handle redirects properly: it remembers the
scheme of the previous request to be able to convert relative `Location`
header to absolute-form before passing it to multi-address client
- `RedirectSingle` rewrites absolute-form `Location` to relative-form
request-target if only relative redirects are allowed;
- Clarify `HttpRequestMetaData.scheme()` contract that it should return
the scheme in lower case;
- Update tests and enhance test coverage;

Result:

`MultiAddressHttpClient` supports only full URLs for request-target,
but sends on the wire relative-form request-target.