Use `PooledByteBufAllocator` for `SslHandler`
Motivation:

`SslHandler` may allocate new direct buffers in case it needs to
cumulate incoming messages or to write the result (encrypted/decrypted)
message. Because we use `ServiceTalkBufferAllocator` when we bootstrap
netty, it propagates this unpooled allocator in the
`ChannelHandlerContext` and `SslHandler` uses it. That causes allocation
of new direct buffers for each request and response, but frequent
allocation of direct memory is costly. We can benefit by using
`PooledByteBufAllocator` inside `SslHandler`. In case of encryption,
newly created pooled buffer will be sent to the network and released by
netty. In case of decryption, we will copy the result into cheap
unpooled heap memory using `CopyByteBufHandler` and release the original
pooled buffer before passing data forward.

Modifications:

- Add `DelegatingChannelHandlerContextWithPooledAllocator` that
wraps the original `ChannelHandlerContext` but overrides `ctx.alloc()`
to return `PooledByteBufAllocator`;
- Add `SslHandlerWithPooledAllocator` that overrides original methods
of `SslHandler` to make sure we wrapped the `ChannelHandlerContext`;
- Change `newHandler`/`newSslHandler` to return
`SslHandlerWithPooledAllocator` before adding it to the pipeline;

Result:

Improved performance when ST reads data for 20-35%, when it sends data
for 25-60%.